Rotate through a series of stats in the city view
The Discord party status line will now rotate through a series of
different city statistics, with the value changing every 30 seconds.
The plugin's previous behavior was to only show the mayor rating.

All numbers shown in the status line are formatted in the US English
locale using SC4's number formatting methods.
The currency values are prefixed with the Simolean/section symbol (§).
The currently used city statistics are:

- Mayor name
- Mayor rating
- Residential population
- Commercial population
- Industrial population
- City age in years
- Monthly net income
- Total funds

This commit also includes a bug fix to update the city name if the user
changes it through a cheat code.